PlatformImpl build
