+++ source/server/BUILD
@@ -184,20 +184,20 @@ envoy_cc_library(
 envoy_cc_library(
     name = "options_lib",
     srcs = ["options_impl.cc"] + select({
-        "//bazel:linux_x86_64": ["options_impl_platform_linux.cc"],
         "//bazel:linux_aarch64": ["options_impl_platform_linux.cc"],
-        "//bazel:linux_ppc": ["options_impl_platform_linux.cc"],
         "//bazel:linux_mips64": ["options_impl_platform_linux.cc"],
+        "//bazel:linux_ppc": ["options_impl_platform_linux.cc"],
+        "//bazel:linux_x86_64": ["options_impl_platform_linux.cc"],
         "//conditions:default": ["options_impl_platform_default.cc"],
     }),
     hdrs = [
         "options_impl.h",
         "options_impl_platform.h",
     ] + select({
-        "//bazel:linux_x86_64": ["options_impl_platform_linux.h"],
         "//bazel:linux_aarch64": ["options_impl_platform_linux.h"],
-        "//bazel:linux_ppc": ["options_impl_platform_linux.h"],
         "//bazel:linux_mips64": ["options_impl_platform_linux.h"],
+        "//bazel:linux_ppc": ["options_impl_platform_linux.h"],
+        "//bazel:linux_x86_64": ["options_impl_platform_linux.h"],
         "//conditions:default": [],
     }),
     # TCLAP command line parser needs this to support int64_t/uint64_t in several build environments.
